Fixed using of CLI options. Add installation for nlohmann::json

+++ src/encora_core/core/utils/Logger.cpp
@@ -1,25 +1,25 @@
 #include <filesystem>
+#include  <iostream>
 
 #include "Logger.h"
+
+#include "spdlog/sinks/ostream_sink.h"
 
 namespace fs = std::filesystem;
 
 namespace EncoraLogger {
-    std::shared_ptr<spdlog::logger> Logger::m_logger = nullptr;
-    bool Logger::m_isInitialized = false;
-
     void Logger::init(const std::string &logDir) {
         if (m_isInitialized) return;
 
         try {
             if (!fs::exists(logDir)) {
-                fs::create_directory(logDir);
+                fs::create_directories(logDir);
             }
 
             // 5MB x 5 files
-            auto fileSink = std::make_shared<spdlog::sinks::rotating_file_sink_mt>(logDir + "/encora.log", 1024 * 1024 * 5, 5);
+            const auto fileSink = std::make_shared<spdlog::sinks::rotating_file_sink_mt>(logDir + "/encora.log", 1024 * 1024 * 5, 5);
             // Colored console sink
-            auto consoleSink = std::make_shared<spdlog::sinks::stdout_color_sink_mt>();
+            const auto consoleSink = std::make_shared<spdlog::sinks::ostream_sink_mt>(std::cout);
             std::vector<spdlog::sink_ptr> sinks { consoleSink, fileSink };
 
             m_logger = std::make_shared<spdlog::logger>("Encora", sinks.begin(), sinks.end());
@@ -27,6 +27,7 @@ namespace EncoraLogger {
 
             m_logger->set_pattern("[%Y-%m-%d %H:%M:%S.%e][%n][%l] %v");
             m_logger->set_level(spdlog::level::debug);
+            m_logger->flush_on(spdlog::level::info);
 
             m_isInitialized = true;
             m_logger->info("Encora logger initialized.");
@@ -36,6 +37,8 @@ namespace EncoraLogger {
     }
 
     void Logger::shutdown() {
+        if (!m_isInitialized) return;
+
         if (m_logger) {
             m_logger->info("Encora logger shutting down.");
             spdlog::drop_all();
@@ -46,13 +49,11 @@ namespace EncoraLogger {
     }
 
     std::shared_ptr<spdlog::logger> &Logger::get() {
-        if (!m_isInitialized) init();
-
         return m_logger;
     }
 
     void Logger::log(const Level level, const std::string &msg) {
-        if (!m_isInitialized) init();
+        if (!m_isInitialized || !m_logger) return;
 
         switch (level) {
             case Level::Trace: m_logger->trace(msg); break;
